classifier/requirements.txt
fastapi
uvicorn
kafka-python
opencv-python-headless 
numpy
pillow
tensorflow  
torch 
torchvision  
